Move local::mutex definitions to source file
